Improve prettyprinting of builtin bound methods.
Previously, a builtin bound method (e.g., `int(1).to_bytes`)
would pretty-print as `None.int.to_bytes # built-in function.`
This is because 1) the `pretty_builtin_method` registered on
BuiltinMethodType would never be called, as BuiltinMethodType is just
an alias for BuiltinFunctionType(!) (`types.BuiltinMethodType is
types.BuiltinFunctionType`), and builtin methods additionally don't set
their `__module__` (but we can retrieve it from `__self__`).

This PR prettyprints `int(1).to_bytes` as
```
int.to_bytes  # built-in bound method
```
it may be even nicer to have it as
```
int(1).to_bytes
```
(i.e. pretty-print the bound value as well) but that's more a general
problem for all bound methods (bound methods of classes written in
Python are not currently prettyprinted), so not in the scope of this PR.